% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.correction.R
\name{background.correction}
\alias{background.correction}
\title{Correct measurements values by background values taken before and/or
after}
\usage{
background.correction(
  x,
  bg,
  t = NULL,
  id = NULL,
  method = c("approx", "mean", "previous", "next"),
  compress = T,
  contact = F,
  id.contact = F,
  tlim = NULL,
  bglim = NULL,
  warn = T
)
}
\arguments{
\item{x}{The measurement and background values, in chronological order of
their measurement.}

\item{bg}{Whether the value is a background measurement (T), or a sample
measurement.}

\item{t}{A numerical value of time.}

\item{id}{Sample id.}

\item{method}{The method for computing the background value at the time of the
sample measurement; "approx" stands for al inear approximation between the
previous and the next background values, using the time of the measurement;
"mean" stands for a mean of the previous and the next background values;
"previous" takes the previous background value; "next", the next one.}

\item{compress}{Whether to remove the background measurements from the output.}

\item{contact}{Whether to check if the background measurements are directly
contiguous with the sample measurements they correct.}

\item{id.contact}{Whether to check that the background measruements used to
correct the sample measurements are directly contiguous with the sample
measurements of identical id they correct.}

\item{tlim}{The maximum time interval allowed for measurements implied in the
same sample correction.}

\item{bglim}{The range of background values allowed. If one value is provided,
bglim = c(-bglim, bglim), otherwise, bglim = range(bglim).}

\item{warn}{Whether to provide a warning when background values exceed given
values, or if the time elapsed between measurement of the sample and of the
background exceeds a given threshold (otherwise this information is only
provided in the output list).}
}
\value{
a list of the initial line of the measurement ($line), the sample id
($id), the sample measurement time ($t), the original sample measurement value
($original.x), the corrected sample measurement value ($corrected.x), the
computed background value used for the correction ($background.x), the time of
measurement of the previous background value ($previous.t), the
background value preceding the sample measurement ($previous.x), the time of
measurement of the next background value ($next.t), and the
background value following the sample measurement ($next.x). Additionally, a
list of the problematic lines of data is provided ($suspect.lines),
indicating the lines of incorrect contact ($suspect.lines$incorrect.contact),
of incorrect id contact ($suspect.lines$incorrect.id.contact), of excessive
background values ($suspect.lines$excessive.background) and of excessive time
(suspect.lines$excessive.time).
}
\description{
Correct measurements values by background values taken before
and/or after, assuming a linear relationship with time. This function can warn
when background values exceed given values, or if the time elapsed between
measurement of the sample and of the background exceeds a given threshold.
}
\examples{
x  <- c(2, 10, 3, 4, 25, 3)
bg <- c(T, F, T, T, F, T)
t  <- c(1:3, 20:22)
id <- c("S1", "S1", "S1", "S2", "S2", "S2")

output <- background.correction(x = x, bg = bg, t = t, id = id)

plot(t, x, pch = 19, ylim = c(0, 30))

lines(t[bg], x[bg])

points(output$t, output$background.x, col = "red", pch = 19)

}
