% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_chris.R
\name{hello_chris}
\alias{hello_chris}
\title{Hello you brilliant people}
\usage{
hello_chris(x = "w")
}
\arguments{
\item{x}{The first letter of your family name}
}
\description{
This should be straightforward, no ?
}
\examples{
hello_chris("z")
}
