% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting.examples.R
\name{inteRmag.examples}
\alias{inteRmag.examples}
\alias{zeq_example}
\alias{frontone14.pmob}
\title{Data for examples}
\description{
Supporting data sets to use in the examples. They will be used in the
examples.
}
\details{
\describe{
  \item{\strong{Oriented data}}{zeq_example}
  \item{\strong{pmob}}{frontone14.pmob}
}
}
