% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.pmob.R
\name{is.pmob}
\alias{is.pmob}
\title{Check if object is a pmob}
\usage{
is.pmob(pmob)
}
\arguments{
\item{pmob}{Object to check.}
}
\description{
Check if object is a pmob (\strong{P}aleo\strong{M}agnetic
\strong{OB}ject).
}
\examples{
pmob <- as.pmob(sampid = "Bosso1", exc = 1, measmin = NA)

is.pmob(pmob)

}
