% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob.col.sort.R
\name{pmob.col.sort}
\alias{pmob.col.sort}
\title{Order the columns of a pmob}
\usage{
pmob.col.sort(pmob)
}
\arguments{
\item{pmob}{Object for which the columns need to be ordered (same order as
the arguments in \code{\link{as.pmob}}.}
}
\description{
Order the columns of a pmob (\strong{P}aleo\strong{M}agnetic
\strong{OB}ject).
}
\examples{
pmob <- as.pmob(sampid = "Bosso1", exc = 1, measmin = NA)
pmob <- pmob[,c(1,5,3,2,4)]

pmob

pmob <- pmob.col.sort(pmob)

pmob

}
