% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob.correct.measure.R
\name{pmob.correct.measure}
\alias{pmob.correct.measure}
\title{Check and normalize measurements values by blank and holder
 measurements}
\usage{
pmob.correct.measure(
  pmob,
  id = "sampid",
  correct.blank = TRUE,
  correct.holder = TRUE,
  compress.blank = FALSE,
  compress.holder = FALSE,
  method.blank = c("mean", "approx", "previous", "next"),
  method.holder = c("previous", "next", "mean", "approx"),
  plot = TRUE,
  time.format = c("day1", "day2", "hour", "min", "sec"),
  center.blank = FALSE,
  center.holder = FALSE,
  contact.blank = FALSE,
  contact.holder = FALSE,
  id.contact.blank = FALSE,
  id.contact.holder = FALSE,
  tlim.blank = NULL,
  tlim.holder = NULL,
  blank.x.lim = NULL,
  blank.y.lim = NULL,
  blank.z.lim = NULL,
  holder.x.lim = NULL,
  holder.y.lim = NULL,
  holder.z.lim = NULL,
  output.suspect.lines = FALSE
)
}
\arguments{
\item{pmob}{a pmob object, for which the measurements hae not been corrected
(for blank and/or holder)}

\item{id}{which column to use as a measurement id (by default, "sampid")}

\item{correct.blank}{whether to correct for blank measurement (TRUE by
default)}

\item{correct.holder}{whether to correct for the holder (TRUE by default);
the measurements should already be corrected for blanks.}

\item{compress.blank}{whether to remove all the blank measurements from the
pmob after correction. This is not advised, to avoid information loss (FALSE
by default).}

\item{compress.holder}{whether to remove all the holder measurements from the
pmob after correction. This is not advised, to avoid information loss (FALSE
by default).}

\item{method.blank}{The methodology used for computing the blank value at
the time of the sample measurement for correction. Supported values are
"approx" for a linear approximation between the previous and the next blank
values, using the time of the measurement, "mean" stands for a mean of the
previous and the next blank values, "previous" takes the previous blank
value, and "next", the next one.}

\item{method.holder}{The methodology used for computing the holder value at
the time of the sample measurement for correction. Supported values are
"approx" for a linear approximation between the previous and the next holder
values, using the time of the measurement, "mean" stands for a mean of the
previous and the next holder values, "previous" takes the previous holder
value, and "next", the next one.}

\item{plot}{whether to plot the behaviors of the blank measurements (x, y
and z) and of the holder measurements (x, y and z).}

\item{time.format}{the format of the plot's time axis; the default,
"day 1" means month-day (the month in abbreviated textual form), "day2"
year-month-day (the month in abbreviated textual form), and "hour", "min",
"sec" the respective time units starting at zero with the first measurement.}

\item{center.blank}{Whether to center the xint, yint and zint of the blank
measurement plots on zero.}

\item{center.holder}{Whether to center the xint, yint and zint of the holder
measurement plots on zero.}

\item{contact.blank}{Whether to check if the blank measurements are
directly contiguous with the sample or holder measurements they
correct.}

\item{contact.holder}{Whether to check if the holder measurements are
directly contiguous with the sample measurements they correct.}

\item{id.contact.blank}{Whether to check that the blank measurements values
used to correct the sample measurements are directly contiguous with the
holder/sample measurements \bold{of identical id} they correct.}

\item{id.contact.holder}{Whether to check that the holder measurements values
used to correct the sample measurements are directly contiguous with the
sample measurements \bold{of identical id} they correct.}

\item{tlim.blank}{The maximum time interval allowed for measurements implied
in the same sample blank correction}

\item{tlim.holder}{The maximum time interval allowed for measurements implied
in the same sample holder correction}

\item{blank.x.lim}{The range of x blank values allowed. This is found in the
plots as red lines. If one value is provided, lim = c(-lim, lim), otherwise,
lim = range(lim).}

\item{blank.y.lim}{The range of y blank values allowed. This is found in the
plots as red lines. If one value is provided, lim = c(-lim, lim), otherwise,
lim = range(lim).}

\item{blank.z.lim}{The range of z blank values allowed. This is found in the
plots as red lines. If one value is provided, lim = c(-lim, lim), otherwise,
lim = range(lim).}

\item{holder.x.lim}{The range of x holder values allowed. This is found in
the plots as red lines. If one value is provided, lim = c(-lim, lim),
otherwise, lim = range(lim).}

\item{holder.y.lim}{The range of y holder values allowed. This is found in
the plots as red lines. If one value is provided, lim = c(-lim, lim),
otherwise, lim = range(lim).}

\item{holder.z.lim}{The range of z holder values allowed. This is found in
the plots as red lines. If one value is provided, lim = c(-lim, lim),
otherwise, lim = range(lim).}

\item{output.suspect.lines}{Whether to have the suspect lines as an output,
rather than the resulting pmob.}
}
\description{
Check and normalize sample (and holder) measurements values by
blank and holder measurements. This function comes with features allowing to
detect abnormalities such as excessive time between sample measurements and
blank/holder measurements, excessive blank holder/values, etc. Problematic
data lines in the pmob can be outputted if needed.
}
\examples{
frontone14.pmob

pmob.correct.measure(frontone14.pmob, plot = F)

}
\seealso{
The \code{\link{background.correction}} function
}
