% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob.merge.R
\name{pmob.merge}
\alias{pmob.merge}
\title{Merge two pmobs together}
\usage{
pmob.merge(pmob1, pmob2, byinitialorder = FALSE)
}
\arguments{
\item{pmob1}{pmob object to complement}

\item{pmob2}{pmob object complementing pmob1}

\item{byinitialorder}{whether to use the \code{initialorder} parameter to
merge the two pmobs. Is set as FALSE as default, as initialorder is
automatically generated for all pmob objects, and should be considered
carefully}
}
\description{
Merge two pmobs (\strong{P}aleo\strong{M}agnetic
\strong{OB}jects) together based on the columns taken account as standard
(see parameters in \code{\link{as.pmob}}) and found in both pmobs. This
function acts like a left join: the first (left) pmob will be completed by
the second (right) pmob.
}
\examples{
pmob1 <- as.pmob(sampid = c("Bosso1","Bosso1", "Bosso2", "Bosso3"),
                 specid = rep("b", 4),
                 slotid = c(1,1,2,3),
                 treattemp = c(20,150, 20, 20))

pmob2 <- as.pmob(sampid = c("Bosso1", "Bosso2", "Bosso3", "Bosso4"),
                 specid = rep("b", 4),
                 slotid = c(1,1,3,4),
                 sampip = c(30, 50, 45, 20))

pmob.merge(pmob1 = pmob1, pmob2 = pmob2)

}
