% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob.na.R
\name{pmob.na}
\alias{pmob.na}
\title{Attribute certain pmob values as NA}
\usage{
pmob.na(pmob, l = list(), remove = FALSE)
}
\arguments{
\item{pmob}{pmob object}

\item{l}{list of parameters and the values that are defined as NA}

\item{remove}{whether to remove the columns made only of NA values (TRUE), or
to keep them (FALSE; is the default)}
}
\description{
Attribute certain values in a pmob
(\strong{P}aleo\strong{M}agnetic \strong{OB}ject) as NA (Not Available), and
allow the columns entirely made of NA values to be removed
}
\examples{
pmob1 <- as.pmob(sampid = c("holder", "Bosso1", "Bosso2", "Bosso3"),
                 specid = c("holder", "b", "c", "NA"),
                 slotid = c("holder", "1", "2", "3"),
                 treattemp = c(20, 20, 20, 20),
                 long = rep(0, 4), lat = rep(0,4))

pmob1

pmob.na(pmob1,
        l = list(specid = c("holder", "NA"),
                 slotid = "holder",
                 long = 0, lat = 0),
        remove = TRUE)

}
