% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob.vol.mass.R
\name{pmob.vol.mass}
\alias{pmob.vol.mass}
\title{Check and fill a pmob for volume- and mass-related unit conversion.}
\usage{
pmob.vol.mass(
  pmob,
  convert.mag = TRUE,
  convert.magsus = TRUE,
  digits = 5,
  vol.main = TRUE,
  default.vol = 10.8 * 1e-06,
  default.mass = 29.16 * 0.001,
  default.dens = 2.7 * 1000,
  vol.lim = 13 * 1e-06,
  mass.lim = NULL,
  dens.lim = NULL
)
}
\arguments{
\item{pmob}{The pmob to treat.}

\item{convert.mag}{Whether to convert the magnetic moment/magnetisation
units.}

\item{convert.magsus}{Whether to convert the magnetic susceptibility units.}

\item{digits}{The number of digits to consider to check the conversion
calculations to already provided values (when the same measurement is
provided in different units).}

\item{vol.main}{In the unique case where only density is available in the
pmob, this parameter details whether the mass will be calculated from the
default volume value (vol.main = TRUE), or on the contrary the volume will
be calculated from the default mass value (vol.main = FALSE).}

\item{default.vol}{The default value of volume. We mention a few standard
values;
11.9 cm3 for the IODP standard (2.5 * 2.5 * 1.9 cm),
10.1 \out{cm<sup>3</sup>} for the Bremen autosampler (2.3 * 2.3 * 1.9 cm),
12.9 \out{cm<sup>3</sup>} for a 1 inch (2.54 cm) slice of a 1 inch-diameter drill core,
10.1 \out{cm<sup>3</sup>} for a 2 cm slice of a 1 inch-diameter drill core.
10.8 \out{cm<sup>3</sup>} for a 2.2 cm slice of a 2.5 cm-diameter drill core (CIMAN-ALP standard, or even more universal ??).
The default is here taken as 10.8 \out{cm<sup>3</sup>}, identical to the .ppl
files (on the arbitrary choice that this is the format that allows to link
inteRmag to PuffinPlot, a WYSIWYG paleomagnetic software that is a good fit
to be combined with our command-line scripting approach). Units are in
\out{m<sup>3</sup>}.}

\item{default.mass}{The default mass. This needs to be coherent with the
default volume and the default density. Units are in kg.}

\item{default.dens}{The default density of the sample, that allows to compute
either mass or volume based on the other. The default value is set as 2.7 g
per cm3, based on the assumed surface density of 2.7 \out{g cm<sup>-3</sup>},
mimicking the density of granite or granodiorite (Williamson & Adams, 1923).
Units are in \out{kg m<sup>-3</sup>}.}

\item{vol.lim}{The maximum volume expected for each sample; any volume value
(calculated or provided) exceeding it triggers a warning. This is set by
default to 13 \out{cm<sup>3</sup>}.}

\item{mass.lim}{The maximum mass expected for each sample; any mass value
(calculated or provided) exceeding it triggers a warning.}

\item{dens.lim}{The range of density expected for each sample. If only one
value is provided, this parameter sets the maximum density value expected for
the sample. Any discrepancy triggers a warning.}
}
\description{
Check and fill a pmob (\strong{P}aleo\strong{M}agnetic
\strong{OB}ject) for volume- and mass-related unit conversion. THIS IS
CURRENTLY ONLY FOR DISCRETE SAMPLES, LONG CORES PARAMETERS ARE NOT TAKEN INTO
ACCOUNT.
}
\details{
If two values are provided for the same magnetisation measurement,
(e.g. if a measurement is provided with both the magnetic moment in Am2 and
in magnetisation by volume), the unnormalised values (xint, yint, zint, and
totmagsus) will be used in priority to compute the missing measurements, and
then the volume-normalised ones (xvol, yvol, zvol, and volmagsus) (in our
example, the mass normalised magnetisation will be obtained from the magnetic
moment). In theory, this should only be relevant in the rounding used on the
values.
}
\examples{
pmob <- as.pmob(sampid = c("A1", "B2", "C3"),
                xint = c(200, 450, NA),
                yint = c(60,  30,  NA),
                zint = c(10,  20,  NA),
                xvol = c(NA, NA, 800000),
                yvol = c(NA, NA,  60000),
                zvol = c(NA, NA,  30000),
                mass = c(0.020, 0.027, 0.010))

pmob.vol.mass(pmob, convert.magsus = F)

}
\references{
\itemize{
  \item Williamson, E. D., and Adams, L. H., 1923. Density distribution in
  the Earth. Journal of the Washington Academy of Sciences 13, n°19, 413-428.
}
}
