% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob_from_AZ.R
\name{pmob_from_AZ}
\alias{pmob_from_AZ}
\title{Convert an .AZ file to a pmob}
\usage{
pmob_from_AZ(file)
}
\arguments{
\item{file}{the directory of the file to convert}
}
\description{
Convert an .AZ file (from the Centro Interuniversitario di
Magnetismo Naturale "Roberto Lanza" - Alpine Laboratory of Paleomagnetism;
CIMAN-ALP format) into a pmob (\strong{P}aleo\strong{M}agnetic
\strong{OB}ject)
}
\examples{
file <- system.file("tests",
                    "CIMAN-ALP_example.AZ",
                    package = "inteRmag")

pmob_from_AZ(file)

}
