% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob_from_CIMAN_ALP.R
\name{pmob_from_CIMAN_ALP}
\alias{pmob_from_CIMAN_ALP}
\title{Convert a CIMAN-ALP .txt data file to a pmob}
\usage{
pmob_from_CIMAN_ALP(file, treat = c("TH", "AF"), units = NULL)
}
\arguments{
\item{file}{the directory of the file to convert}

\item{treat}{the treatment used for samples: either thermal ("TH") or
alternating field demagnetization ("AF")}

\item{units}{the units of the treatment: can be "Celsius" (default for
thermal demagnetization), "Kelvin",  "Millitesla" (default for alternating
field demagnetization) or "Tesla".}
}
\description{
Convert a CIMAN-ALP .txt data file (from the Centro
Interuniversitario di Magnetismo Naturale "Roberto Lanza" - Alpine
Laboratory of Paleomagnetism; CIMAN-ALP format) into a pmob
(\strong{P}aleo\strong{M}agnetic \strong{OB}ject)
}
\examples{
file <- system.file("tests",
                    "CIMAN-ALP_example.txt",
                    package = "inteRmag")

pmob_from_CIMAN_ALP(file, treat = "TH", units = "Celsius")

}
