% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob_from_ppl.R
\name{pmob_from_ppl}
\alias{pmob_from_ppl}
\title{Convert an .ppl file to a pmob}
\usage{
pmob_from_ppl(
  file,
  depth.factor = 1,
  time.format = "\%m/\%d/\%y \%H\%M",
  magsus.type = c("vol", "mass", "tot"),
  magsus.factor = 1,
  irm.axis = c("x", "y", "z", "-x", "-y", "-z"),
  arm.bias.axis = c("x", "y", "z", "-x", "-y", "-z"),
  defaults = list(DISCRETE_ID = "UNSET", DEPTH = "null", RUN_NUMBER = "-1", TIMESTAMP =
    "UNSET", SLOT_NUMBER = "-1", MEAS_TYPE = "UNSET", X_MOMENT = NULL, Y_MOMENT = NULL,
    Z_MOMENT = NULL, MAG_SUS = "NaN", VOLUME = "10.8", AREA = "4.0", SAMPLE_AZ = "NaN",
    SAMPLE_DIP = "NaN", FORM_AZ = "NaN", FORM_DIP = "NaN", MAG_DEV = "0.0", TREATMENT =
    "UNKNOWN", AF_X = "NaN", AF_Y = "NaN", AF_Z = "NaN", TEMPERATURE = "NaN", IRM_FIELD =
    "NaN", ARM_FIELD = "NaN", ARM_AXIS = "UNKNOWN", PP_SELECTED = NULL, PP_ANCHOR_PCA =
    NULL,      PP_HIDDEN = NULL, PP_ONCIRCLE = NULL, PP_INPCA = NULL)
)
}
\arguments{
\item{file}{the directory of the file to convert}

\item{depth.factor}{the factor by which the depth values in the .ppl file
need to be multiplied to convert to meters (e.g. if the units are
centimeters, the depth.factor parameter needs to set at 100). By default the
function assumes that the units are in meters (depth.factor = 1).}

\item{time.format}{how the time in the TIMESTAMP column of the .ppl file is
formatted. For more details, check the \code{strptime} function in R base.}

\item{magsus.type}{whether the magnetic susceptibility units are normalised
by volume (magsus.type = "vol"), by mass (magsus.type = "mass"), or are not
normalised (i.e. total magnetic susceptibility; magsus.type = "tot"). By
default the units are considered to be normalised by volume.}

\item{magsus.factor}{the factor by which the magnetic susceptibility values
in the .ppl file need to be multiplied to convert to the proper magnetic
susceptibility units. If the magnetic susceptibility is normalised by volume,
the values need to be converted into dimensionless SI units, if normalised
by mass, into cubic meters by kilograms, and if not normalised (total
magnetic susceptibility), into cubic meters. By default magsus.factor is set
to 1.}

\item{irm.axis}{the axis along which the Isothermal Remanent Magnetisation
(IRM) was applied (can be "x", "y", "z", "-x", "-y" or "-z")}

\item{arm.bias.axis}{the axis along which the bias field of the Anhysteretic
Remanent Magnetisation (ARM) was applied (can be "x", "y", "z", "-x", "-y" or
"-z")}

\item{defaults}{a list of the default values of the columns in the .ppl file
(they will be translated to NA values; or in the case of the volume and area,
they will be signaled as being default/not exact values).}
}
\description{
Convert .ppl file (from the PuffinPlot format) into a pmob
(\strong{P}aleo\strong{M}agnetic \strong{OB}ject)
}
\examples{
file <- system.file("tests",
                    "PuffinPlot_example.ppl",
                    package = "inteRmag")

pmob_from_ppl(file)

}
