% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmob_to_ppl.R
\name{pmob_to_ppl}
\alias{pmob_to_ppl}
\title{Convert a pmob to an .ppl file}
\usage{
pmob_to_ppl(
  pmob,
  file,
  time.format = "\%m/\%d/\%y \%H\%M",
  magsus.type = c("none", "vol", "mass", "tot"),
  other.columns = TRUE
)
}
\arguments{
\item{pmob}{the data in \strong{P}aleo\strong{M}agnetic \strong{OB}ject form
to convert into a .ppl file}

\item{file}{the name of the .ppl file}

\item{time.format}{how the time in the TIMESTAMP column of the .ppl file
needs to be formatted. For more details, check the \code{strptime} function
in R base.}

\item{magsus.type}{whether the magnetic susceptibility units need to be the
ones normalised by volume (magsus.type = "vol"), by mass (magsus.type =
"mass"), or are not normalised (i.e. total magnetic susceptibility;
magsus.type = "tot"). By default the units are considered to be normalised
by volume.}

\item{other.columns}{if TRUE, all columns that have headers that are not
regular pmob or .ppl ones will be added into the .ppl file as flags (if
the values are "false" or "true" or as notes if the values are characters).
The parameter can also be a vector of all the column names to use as flags or
notes.}
}
\description{
Convert a pmob (\strong{P}aleo\strong{M}agnetic \strong{OB}ject)
into a.ppl file (from the PuffinPlot format).
}
\examples{
file <- system.file("tests",
                    "PuffinPlot_example.ppl",
                    package = "inteRmag")

pmob_Bosso <- pmob_from_ppl(file)

\dontrun{
pmob_to_ppl(pmob_Bosso, "Bosso_Example")}

}
