% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ppl.R
\name{read.ppl}
\alias{read.ppl}
\title{Read .ppl files}
\usage{
read.ppl(file)
}
\arguments{
\item{file}{the directory of the file to convert}
}
\description{
Convert a .ppl file (from the PuffinPlot format) into a table
}
\examples{
file <- system.file("tests",
                    "PuffinPlot_example.ppl",
                    package = "inteRmag")

read.ppl(file)

}
\references{
Lurcock, P. C., and G. S., Wilson. 2012. "PuffinPlot: A versatile,
user-friendly program for paleomagnetic analysis". Geochemistry, Geophysics,
Geosystems, 13, Q06Z45, https://doi.org/10.1029/2012GC004098.
}
