% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.ppl.R
\name{write.ppl}
\alias{write.ppl}
\title{Write .ppl files}
\usage{
write.ppl(x, file = "")
}
\arguments{
\item{x}{the table to convert.}

\item{file}{the directory of the file to convert}
}
\description{
Convert a table into a .ppl file (from the PuffinPlot format)
}
\examples{
\dontrun{
write.ppl(data.frame(), "empty_file")}

}
\references{
Lurcock, P. C., and G. S., Wilson. 2012. "PuffinPlot: A versatile,
user-friendly program for paleomagnetic analysis". Geochemistry, Geophysics,
Geosystems, 13, Q06Z45, https://doi.org/10.1029/2012GC004098.
}
